White List & Cubic Chunk update
Adds a whitelist of blocks the mod will ignore.  It will not make them
slower or more exhausting to dig.

Support Cubic Chunks.   It will default to be as hard as it would at
depth 5.   If you have downward digging active, it will continue to
work.

Adds version checking.

Fixes a design flaw/bug with axes and shovels.   They were relying on
"isharvestable()" which wasn't working for them.